Updated Code with Home Screen, Sign In Page, Register Page, Hotel Details Displayed on home screen in tabular format from Database

+++ hotelreservationsystem/src/main/java/edu/depaul/cdm/se452/teamnosleep/hotelreservationsystem/Repositories/HotelsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import edu.depaul.cdm.se452.teamnosleep.hotelreservationsystem.entities.Hotels;
 
 
-public interface HotelsRepository extends JpaRepository<Hotels, Long>  {
+public interface HotelsRepository extends JpaRepository<Hotels, Integer>  {
     List<Hotels> findByLocationId(Integer id);
 }
