+++ hotelreservationsystem/src/main/java/edu/depaul/cdm/se452/teamnosleep/hotelreservationsystem/Repositories/LocationsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import edu.depaul.cdm.se452.teamnosleep.hotelreservationsystem.entities.Locations;
 
 
-public interface LocationsRepository extends JpaRepository<Locations, Long>  {
+public interface LocationsRepository extends JpaRepository<Locations, Integer>  {
     Locations findByLocationName(String location);
 }
